Updated code structure
