+++ requirements-dev.txt
@@ -1,4 +1,2 @@
 pytest==8.1.1
-build==1.2.1
-twine==5.0.0
