+++ src/deconfig/__version__.py
@@ -1,0 +1,5 @@
+__title__ = "deconfig"
+__description__ = "Decorator driven Configuration Library"
+__version__ = "0.1.0"
+__author__ = "ybhaw"
+__license__ = "MIT"
