--- src/deconfig/core/adapter/adapter_error.py
@@ -1,5 +1,0 @@
-class AdapterError(Exception):
-    """
-    Raised when adapter fails to get field value.
-    """
-    pass
